hot-rank-web/requirements-utf8.txt
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
beautifulsoup4==4.13.3
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
cssselect==1.2.0
Deprecated==1.2.18
dnspython==2.7.0
email-validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.8
feedgen==1.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
json-repair==0.39.1
limits==4.0.1
lxml==5.3.1
markdownify==1.0.0
motor==3.7.0
packaging==24.2
psycopg2-binary==2.9.10
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0
wrapt==1.17.2
aiohttp
